pyproject.toml
[project]
name = "chemdraw-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.10.6",
    "loguru>=0.7.3",
    "mcp[cli]>=1.12.2",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "rdkit>=2025.3.3",
    "uvicorn-loguru-integration>=0.3.1",
]
